Specify default values of TARGMAX
a) Specify the meaning of the absence of TARGMAX -> clients shouldn't assume they
   are greater than 1
b) Specify the meaning of missing commands -> also 1.

At least InspIRCd is in case a). I don't know about any IRCd in case b),
but since we don't have any defined list of commands, I think it just
makes sense to include it.