Specify WHOWAS
The syntax and base description are copied from
https://datatracker.ietf.org/doc/html/rfc1459#section-4.5.3
with two changes:

1. removed `<server>` because it is pointless
2. "`<count>` SHOULD be a non-positive number" because one
   implementation (ircu2) does not handle `0` the same way others do.
   (besides, there is no reason for clients to send that)

I ignored https://datatracker.ietf.org/doc/html/rfc2812#section-3.6.3
because noone seems to implement wildcards, and the only implementation
supporting TARGMAX WHOWAS (Bahamut) does not follow the spec (it returns
replies in query order instead of chronological order)

I added a description of replies, that matches every implementation I
could find.
In particular, I made it a requirement that `RPL_WHOWASUSER` must come
first, because otherwise it would be hell for clients to parse this.

Finally, I made it explicit that `RPL_ENDOFWHOWAS` should not be used
when the nick is missing, to match existing implementations (as opposed
to what the RFCs say).