Clarify ELIST C and T
The phrasing before this commit was an attempt at clarifying
https://datatracker.ietf.org/doc/html/draft-hardy-irc-isupport-00#section-4.8
which was very ambiguous (didn't even mention minutes or being a time
delta...).

This matches the behavior of Solanum/Charybdis, InspIRCd, and ircu2 (and
Plexus4/Hybrid on `T` only).

The "widespread bug" is in UnrealIRCd (and Plexus4/Hybrid on `C` only).