Change recommended handling of missing CHANTYPES (#230)
* Change recommended handling of missing CHANTYPES

Formerly we recommended treating missing CHANTYPES as an empty string,
i.e. no channel types. This switches to the recommendation in
https://datatracker.ietf.org/doc/draft-brocklesby-irc-isupport/
to assume the RFC1459 default of `#&`.

* distinguish between missing value and missing token

Feedback from #ircdocs: `CHANTYPES` or `CHANTYPES=` should be respected as
explicitly indicating that no channel types are supported (contra the
Brocklesby draft), but if the parameter is missing entirely, we should treat
it as `CHANTYPES=#&`