RPL_WHOISCHANNELS: Mention trailing space
RFCs 1459 and 2812 *require* this space to be present, and some servers
do send it (irc2, ircu2, charybdis, unrealircd).

I am keeping it out of the grammar because it looks like an accident in
implementations that somehow made it to the RFCs; and most IRCds corrected
it already (Ergo, Hybrid, InspIRCd, ngIRCd, Plexus4, Solanum).